requirements.txt
Django>=2.2
djangorestframework>=3.9.0
PyNaCl>=1.3.0
PyJWT>=1.7
cryptography>=2.6.0
